client/requirements.txt
absl-py
pytest
requests
typing